requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-decouple==3.8
requests==2.31.0
celery==5.3.1
redis==4.6.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0
django-filter==23.3
gunicorn==21.2.0
whitenoise==6.6.0
reportlab==4.0.7
Pillow==10.1.0
xlwt==1.3.0
qrcode==7.4.2
openpyxl==3.1.2
matplotlib==3.8.2
pandas==2.1.4